Fix Chaperon. Revert Monitor.
Reverting Monitor to use program. It will always be the case that you'll
use the command line to invoke the program. We're not going to assume
that we know how to invoke your program for replay. You might have to
start it with stub networking. It is probably not on the same path on
your development workstation as it is in your production environment.

Trying to preserve the startup information is also difficult because we
construct the Paxos participant prior to reading the file. If we do want
to save that information we're going to have to work on Channel for a
while. It is not a trivial change.

Chaperon was not working correctly with the new Kibitz. Adjusted for new
bootstrap and join.