Schedule and reschedule failure detection.
Schedule and reschedule failure detection. When a Paxos object notifies
Compassion that it has received a message during healthy communication,
we push back failure detection. If we don't get a healthy ping for a
while we start the failure process of the Chaperon.

Setting the scheduled event for a specific key replaces the event
scheduled for that key previously. This means that we stop our
embarkation events the moment we starting getting healthy pings.

Embarkation is the process of telling the leader we want to join. We
repeatedly submit embarkation requests which are unidirectional. We only
know if it succeeds if our Paxos object starts emitting healthy pings.

Our embarkation logic will detect whether or not we should give up on
embarkation. It will detect whether the "republic" we're trying to join
has failed.