Start republic collapse detection.
Dialed back the republic collapse detection so that it will fire during
our tests. It is going to determine that the republic has not collapsed
and schedule a retry for itself.

You can see the republic collapse detection logic in recoverable. It
checks to see that there is no greater republic, because that would mean
that the current republic no longer exists. It checks to see that a
majority of the members of the government are present.

Checked that everything is going cancel correctly. The only asynchronous
operation is the request to participate sent to the leader. That request
is designed to be made repeatedly until it succeeds, to be
inconsequential if it is duplicated, in-process or the effort abandoned.