Remove Chaperon race condition.
While waiting for the census, the colleague could be destroyed and
replaced so that the island and id of the event envelope retrieves a
colleague subsequent instance of that application.

We get the colleague immediately after being called by the timer, gather
the census, then check to see if that particular colleague has been
destroyed before proceeding.