Separate postbacks so we can add keep-alive ping.
Moving postbacks to a common user agent so I can add a Happenstance
scheduler to schedule keep-alive pings. The keep-alive pings will occur
at a specified interval. When there is a postback I want to push the
ping next back by the specified interval. No ping if we're chatting with
the application anyway.