Revert "Separate postbacks…"
This reverts commit 0ee8b8426637ef66ff6ae274b0bc9172f38295ba.

Not going to reset ping after a postback. Ping is going to be separate,
consistant, and it will run during play and replay. Weaving too many
concerns in this way. Conference needs to know about the ping to reset
it.

It might reset it when it has been canceled by a termination, so now it
would have to check to see if it is allowed to reset it.

You see, currently the response to a log entry could play through even
though the colleague instance has been destroyed. It is benign. If the
application wants to make requests during the log entry, it will find
that its token is rejected. We're not going to resurrect a dead
colleague.

However, we might resurrect a dead keep-alive ping if we simply reset
the ping timer after a successful postback. Now the ping timer and
postback logic has to be aware of the colleague object.