Add keep-alive ping.
Add a keep-alive ping. The application should maintain a shutdown timer
and shutdown when not pinged within an interval of the application's
choosing. By default pings will arrive once every second. No ping means
that Compassion has crash restarted or has otherwise decided to kill
that participant.

As noted elsewhere, the keep-alive ping is a network operation, not
part of the application's transaction log. Regular operation and
debugging replay will not replay the same pings in the same order. Pings
occur at a regular interval even if the application is receiving
postbacks.