Refactor broadcast message formats.
Update primarily so I can have a snapshot of the islander that initiated
the broadcast that includes both id and arrival promise because that
islander may depart before the broadcast is reduced.

Otherwise, flattened reduced properties in the HTTP postback,
eliminating the nested `body` object. While grouping properties in
a `request` and `reduced` envelopes sent through Paxos.