Unit test population census race.
Unit test a race condition where we request a census in order to
determine a chaperon action for a colleague, but the colleague is
destroyed while we're waiting for the result. Note that race condition
does not mean "bug." We've carefully considered our chaperon logic
so that the population census is the only asynchronous request that can
return to an altered application state. All other operations are either
synchronous or their asynchronous operations are requests that are
benign if the sidecar application has crashed.