Destroy constituent when any are unrecoverable.
Destroy the entire constituent when any become unrecoverable. We are now
an Olio constituent. We are part of a multi-process program where all of
the processes are long lived. There is no more ping. There is no good
way for a sibling that is dependent on us to restart. Even if only one
island of many is unrecoverable we need to terminate to allow the
program to restart for the dependent upon the unrecoverable to reconnect
and reconfigure.