Upgrade `staccato` to 8.0.x.
Staccato Readable no longer raises an error if it receives an error for
the `"error"` event of the underlying stream. To restore this behavior
you can invoke `Readable.raise()` to raise any error it might have
intercepted. `Readable` did a pretty good job of intercepting errors.
Writable didn't so we're moving away from the convert-event-stream-
errors-to-exceptions model in Staccato. Not moving away from it here,
though.